requirements.txt
asgiref==3.7.2
bleach==6.1.0
certifi==2023.7.22
charset-normalizer==3.3.2
django>=4.0.0,<4.1.0
django-braces==1.15.0
django-cors-headers==4.3.1
django-filter==23.3
django-summernote==0.8.20.0
djangorestframework==3.14.0
idna==3.4
Markdown==3.5.1
Pillow==10.1.0
psycopg==3.1.12
psycopg-binary==3.1.12
psycopg2-binary==2.9.9
pytz==2023.3.post1
PyYAML==6.0.1
requests==2.31.0
six==1.16.0
sqlparse==0.4.4
typing_extensions==4.8.0
urllib3==2.1.0
webencodings==0.5.1
